requirements.txt
google-adk
google-generativeai
python-dotenv
Pillow
opencv-python
numpy
fastapi
uvicorn